Moved code to CoffeeProcessor
